rust: Finish splitting up websocket.rs
